+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -15,8 +15,10 @@ public class ItemDto {
     private long id;
     @NotBlank(groups = Create.class)
     private String name;
+
     @NotBlank(groups = Create.class)
     private String description;
+
     @NotNull(groups = Create.class)
     private Boolean available;
 }
