+++ src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -20,7 +20,7 @@ import ru.practicum.shareit.user.User;
 public class Item {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
