+++ src/main/java/ru/practicum/shareit/user/User.java
@@ -23,6 +23,6 @@ public class User {
 
     @Email
     @NotBlank(groups = Create.class)
-    @Column(name = "email")
+    @Column(name = "email", unique = true)
     private String email;
 }
